src/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2==2.9.10
PyJWT==2.9.0
redis==5.2.1
sqlparse==0.5.3
